requirements.txt
asgiref==3.5.2
bleach==5.0.1
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-multiselectfield==0.1.12
django-summernote==0.8.20.0
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
webencodings==0.5.1
zipp==3.8.1
